% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmert.R
\name{lmer_t}
\alias{lmer_t}
\title{Inference based on a t-distribution for 'lmer' objects.}
\usage{
lmer_t(object, method = c("Heuristic", "Satterthwaite", "Kenward-Roger"),
  level = 0.95)
}
\arguments{
\item{object}{an object representing a linear mixed effects model, as returned by \code{lmer}}

\item{method}{a character string that selects the method used to determine the degrees of freedom.}

\item{level}{a number between 0 and 1, the level of the confidence intervals}
}
\value{
The function returns a list with two elements \code{"lmer"} - the original
object - and \code{"coefTab"} which is a matrix with estimates, standard errors,
t-statistics, degrees of freedom, lower and upper confidence intervals, and p-values.
}
\description{
Function \code{lmer_t} provides confidence intervals and p-values
based on the m-l-1 heuristic, the Satterthwaite and Kenward-Roger methods for models fitted with \code{lmer}.
}
